#10 Split out GeneralConfigTab into subcomponents
